client/requirements.txt
Flask==2.3.3
python-telegram-bot==20.0
requests==2.31.0
